+++ packages/json-schema-to-zod/src/json-schema-to-zod.ts
@@ -3,13 +3,13 @@ import type { z } from 'zod';
 import { parseSchema } from './parsers/parse-schema';
 import type { JsonSchemaToZodOptions, JsonSchema } from './types';
 
-export const jsonSchemaToZod = <T extends z.ZodTypeAny = z.ZodTypeAny>(
-	schema: JsonSchema,
-	options: JsonSchemaToZodOptions = {},
-): T => {
-	return parseSchema(schema, {
-		path: [],
-		seen: new Map(),
-		...options,
-	}) as T;
+export const jsonSchemaToZod = (
+  schema: JsonSchema,
+  options: JsonSchemaToZodOptions = {}
+): z.ZodType => {
+  return parseSchema(schema, {
+    path: [],
+    seen: new Map(),
+    ...options,
+  });
 };