+++ packages/json-schema-to-zod/src/types.ts
@@ -1,4 +1,5 @@
 import type { ZodTypeAny } from 'zod';
+import type { JSONSchema7 } from 'json-schema';
 
 export type Serializable =
   | { [key: string]: Serializable }
@@ -9,64 +10,7 @@ export type Serializable =
   | null;
 
 export type JsonSchema = JsonSchemaObject | boolean;
-export type JsonSchemaObject = {
-  // left permissive by design
-  type?: string | string[];
-
-  // object
-  properties?: { [key: string]: JsonSchema };
-  additionalProperties?: JsonSchema;
-  unevaluatedProperties?: JsonSchema;
-  patternProperties?: { [key: string]: JsonSchema };
-  minProperties?: number;
-  maxProperties?: number;
-  required?: string[] | boolean;
-  propertyNames?: JsonSchema;
-
-  // array
-  items?: JsonSchema | JsonSchema[];
-  additionalItems?: JsonSchema;
-  minItems?: number;
-  maxItems?: number;
-  uniqueItems?: boolean;
-
-  // string
-  minLength?: number;
-  maxLength?: number;
-  pattern?: string;
-  format?: string;
-
-  // number
-  minimum?: number;
-  maximum?: number;
-  exclusiveMinimum?: number | boolean;
-  exclusiveMaximum?: number | boolean;
-  multipleOf?: number;
-
-  // unions
-  anyOf?: JsonSchema[];
-  allOf?: JsonSchema[];
-  oneOf?: JsonSchema[];
-
-  if?: JsonSchema;
-  then?: JsonSchema;
-  else?: JsonSchema;
-
-  // shared
-  const?: Serializable;
-  enum?: Serializable[];
-
-  errorMessage?: { [key: string]: string | undefined };
-
-  description?: string;
-  title?: string;
-  default?: Serializable;
-  readOnly?: boolean;
-  not?: JsonSchema;
-  contentEncoding?: string;
-  nullable?: boolean;
-};
-
+export type JsonSchemaObject = JSONSchema7;
 export type ParserSelector = (schema: JsonSchemaObject, refs: Refs) => ZodTypeAny;
 export type ParserOverride = (schema: JsonSchemaObject, refs: Refs) => ZodTypeAny | undefined;