+++ packages/json-schema-to-zod/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @composio/json-schema-to-zod
 
+## 0.1.6
+
+### Patch Changes
+
+- 37a1f01: Feat better connected account creation flow
+
+## 0.1.5
+
+### Patch Changes
+
+- df31cc2: Fix json schema parsing
+
 ## 0.1.2
 
 ### Patch Changes